Udpated index.js for seeds and renamed a seeding file
